+++ croud/__main__.py
@@ -54,7 +54,12 @@ from croud.organizations.users.commands import (
 )
 from croud.parser import Argument, create_parser
 from croud.products.commands import products_list
-from croud.projects.commands import project_create, project_delete, projects_list
+from croud.projects.commands import (
+    project_create,
+    project_delete,
+    project_edit,
+    projects_list,
+)
 from croud.projects.users.commands import (
     project_users_add,
     project_users_list,
@@ -288,6 +293,20 @@ command_tree = {
                     Argument("-y", "--yes", action="store_true", default=False),
                 ],
                 "resolver": project_delete,
+            },
+            "edit": {
+                "help": "Edit the specified project.",
+                "extra_args": [
+                    Argument(
+                        "-p", "--project-id", type=str, required=True,
+                        help="The project ID to use.",
+                    ),
+                    Argument(
+                        "--name", type=str, required=False,
+                        help="The new project name to use.",
+                    ),
+                ],
+                "resolver": project_edit,
             },
             "list": {
                 "help": (